feat: PCA reintegration with segmentation algorithm [RQ-0301]
