+++ released_plugins/v3d_plugins/MIND/object_script.soma_segmentation.Release
@@ -1,5 +1,6 @@
 release/soma_segmentation_plugin.o
 release/ResolutionDialog.o
+release/pcaAnalysis.o
 release/v3d_message.o
 release/v_neuronswc.o
 release/basic_surf_objs.o
