+++ released_plugins/v3d_plugins/MIND/soma_segmentation_plugin.h
@@ -9,11 +9,10 @@
 
 #include <v3d_interface.h>
 
-#include <QFileDialog>  // Add this include
 #include <QtGui>
-#include <fstream>  // Add this for file output
 
 #include "cellSegmentation_plugin.h"
+#include "pcaAnalysis.h"
 
 // A basic structure for a marker
 struct MyMarker {
@@ -64,23 +63,9 @@ MIND_4DImage *reconstruction_func(V3DPluginCallback2 &callback, QWidget *parent,
 void isotropic_correction_func(V3DPluginCallback2 &callback, QWidget *parent,
                                input_PARA &PARA, bool bmenu);
 
-// PCA Analysis
+// PCA analysis function
 
 void pca_func(V3DPluginCallback2 &callback, QWidget *parent, input_PARA &PARA,
               bool bmenu);
-
-void analyzeSomaPCA(unsigned char *labeledData, V3DLONG N, V3DLONG M, V3DLONG P,
-                    const LocationSimple &lm, int somaIndex);
-
-void savePCAResultsToCSV(
-    const QString &filename, int somaIndex, const LocationSimple &lm,
-    double pc1, double pc2, double pc3, const double *vec1, const double *vec2,
-    const double *vec3, double x_center, double y_center, double z_center,
-    QWidget *parent = nullptr,
-    bool *saveEnabled = nullptr);  // Add save flag parameter
-
-void drawLine(Image4DSimple *image, double *from, double *to);
-
-void visualizePCA_func(V3DPluginCallback2 &callback, QWidget *parent);
 
 #endif  // __SOMA_SEGMENTATION_PLUGIN_H__
